server/requirements.txt
# https://pypi.org/project/flask
Flask==3.0.2

# https://pypi.org/project/gunicorn
gunicorn==21.2.0

Pillow==10.0.1

google-cloud-datastore

google-cloud-storage